requirements.txt
torch==1.7.1
torchvision==0.8.2
torchaudio==0.7.2
tensorboard==2.5.0
pillow==8.2.0
numpy>=1.18
scipy>=1.5
scikit-learn==0.24.0
tqdm==4.56.0
h5py==3.1.0
matplotlib==3.3.3
pandas==1.2.0
opencv-python>=4.1
seaborn==0.11.1
netcal==1.1.3
